Update sphere_fibonacci_grid_points.m
The original code produced points that were actually not uniformly distributed on a sphere. The points were on clearly distinguishable spirals with larger gaps between them. Run both the new and old code with 5001 points, plot them and compare to see what I mean. I have not numerically checked whether my code is correct, but it directly implements the equations in the paper and the resulting points look exactly like the example figures in the paper.
Also, I failed to find an e-mail address of John Burkardt, hence I did not contact him in this issue.